task_manager/requirements.txt
altair==5.4.1
arrow==1.3.0
asgiref==3.8.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
crispy-bootstrap4==2024.1
Django==5.1.7
django-crispy-forms==2.1
djangorestframework==3.15.2
docopt==0.6.2
dotenv==0.0.5
extra-streamlit-components==0.1.71
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isoduration==20.11.0
Jinja2==3.1.4
Js2Py==0.74
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_server==2.27.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.13.4
numpy==2.1.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pipwin==0.5.2
protobuf==5.28.3
pyarrow==18.0.0
PyAudio==0.2.14
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.18.0
pyjsparser==2.7.1
PyJWT==2.9.0
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
SpeechRecognition==3.11.0
sqlparse==0.5.3
streamlit==1.40.1
streamlit-audiorecorder==0.0.5
streamlit-authenticator==0.4.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
uri-template==1.3.0
urllib3==2.2.3
watchdog==6.0.0
webcolors==24.8.0
